% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cenWeight.R
\name{cenWeight}
\alias{cenWeight}
\title{Function to estimate Efforn weights}
\usage{
cenWeight(x, n, dist = NULL)
}
\arguments{
\item{x}{an identifier specifying which observations are censorted}

\item{n}{an interger specifying the total number of observations}

\item{dist}{an optional covairate of lenth n specifiyin and additional
distance, defualts to 'NULL' and all observations are given a nominal
distance of 1}
}
\value{
A set of weights for use in the kernel hazard estimation
}
\description{
This function will calculate a set of weights to be used in the kernel
density estimation of a hazard function
}
\details{
This performs the procedure which redistributes the weights of
censored observations to all observations remaining at risk.  Where dist is
not spefied this will provide a set of weights which will produce an
esitmator close to the Kaplan Meier estimator.  Where a distance is applied a
set of weights is produced which can adjust the density estiamte.
}
