% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didi.object.R
\docType{data}
\name{didi.object}
\alias{didi.object}
\title{Fitted \code{didi} object}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
didi.object
}
\description{
An object returned by the \code{didEst} function, inheriting from class
\code{didi} and representing a didi Estimator object.
}
\author{
Richard Jasckson (\email{richj23@liverpool.ac.uk})
}
\keyword{datasets}
