% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didiEst.R
\name{didiEst}
\alias{didiEst}
\title{Function to perform kernel estimation using normally distributed kernels.}
\usage{
didiEst(st, cen, x, bw = NULL, bw.x = NULL, type = "survival", size = 100)
}
\arguments{
\item{st}{a vector of survival times}

\item{cen}{a vector of censoring indicators}

\item{x}{a vector contianing a continuous covariate}

\item{bw}{a bandwidth parameter for time}

\item{bw.x}{a bandwidth parameter for the covariate}

\item{type}{type of estimator to be provided either cumulative hazard or
survival, defaults to survival}

\item{size}{a parameter setting the number of points at which the kernals
should be evaluated.  Default to 100.}
}
\value{
A object of class "didi" containing
* time: a vector specifying at which times the estimator was evaluated
* cov: a vector specifying at which covariate values the estimator was evaluated
* xMatrix: a matrix supplying the double weighted kernel estimators
}
\description{
This function to produce the 'Didi' estimator.  A double weighted kernel
estimator of the survival function in the presence of a continuous covariate.
}
\details{
This performs the procedure which estimates the kernel density of a
set of data points, x at a set of evaluation points y.  The function creates
a grid with of equal points of length 'size' between 0 and the maximum
observed event time.  The kernel estimator is then evaluated at each unique
covariate value
}
