% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didiPlot.R
\name{didiPlot}
\alias{didiPlot}
\title{Function to plot the Didi estimator.}
\usage{
didiPlot(x)
}
\arguments{
\item{x}{an object of class "didi"}
}
\value{
A fileld contour plot for the didi estimator
}
\description{
This function to plot the 'Didi' estimator.  This isuses the 'filled.contour'
opint in base graphics and is now deprecited hving been replaced by
plot.didi() which makes use of ggplot2.
}
\details{
This takes the inputs of the "didi" object and plots them using a
filled contour plot.
}
