% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernEst.R
\name{kernEst}
\alias{kernEst}
\title{Function to perform kernel estimation using normally distributed kernels.}
\usage{
kernEst(x, y, bw)
}
\arguments{
\item{x}{vector of covariate values}

\item{y}{a optional vector specifying where the kernels should be evaluated}

\item{bw}{a bandwidth parameter specify the standard deviation of the kernels}
}
\value{
A matrix of dimension x by y
}
\description{
This function performs the kernel estimation procedure for a set of values x,
at specified points y with a bandwidth bw.  For use with the didiEst function.
}
\details{
This performs the procedure which estimates the kernel density of a
set of data points, x at a set of evaluation points y.
}
