% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.didi.R
\name{plot.didi}
\alias{plot.didi}
\title{Function to perform kernel estimation using normally distributed kernels.}
\usage{
\method{plot}{didi}(x, ylab = "cov", xlab = "time", ...)
}
\arguments{
\item{x}{an object of class "didi"}

\item{ylab}{a text input fo the y axis}

\item{xlab}{a text input fo the x axis}

\item{...}{not used}
}
\value{
A ggplot object giving the filled contour plot
}
\description{
This function to produce the 'Didi' estimator.  A double weighted kernel
estimator of the survival function in the presence of a continuous covariate.
}
\details{
This takes the inputs of the "didi" object and plots them using a
filled contour plot.  This makes use of the stat_contour_filled option within
ggplot 2
}
